+++ src/features/jira/projects/repositories/project-permission.repository.ts
@@ -33,7 +33,7 @@ export class ProjectPermissionRepositoryImpl
     });
 
     const permissions = await this.httpClient.sendRequest<ProjectPermissions>({
-      endpoint: "user/permission/search",
+      endpoint: "mypermissions",
       method: "GET",
       queryParams: {
         projectKey,